Update Sidechain-SDK-extension.rst
I think I forgot to mention: there are several instances where a pair of back-ticks - ` - is seemingly used in place of a double quote, ". There might be another explanation: perhaps it's a kind of escape that allows the use of the back-tick as a punctuation symbol, or maybe some kind of format indicator for this editor. If a double-quote IS meant, perhaps a global replace of `` with " should be done.